JAVA-24: Move default handler into its own class.
